Avoid python 2.5 requirement by using int instead of Py_ssize_t (transplanted from 7e06ee4ecabcfcd79a79c063d77b404962836c74)

+++ python/rpmmodule.c
@@ -63,7 +63,7 @@ static PyObject * archScore(PyObject * self, PyObject * args, PyObject * kwds)
 static PyObject * signalsCaught(PyObject * self, PyObject * check)
 {
     PyObject *caught, *o;
-    Py_ssize_t llen;
+    int llen;
     int signum, i;
     sigset_t newMask, oldMask;
